Remove dynamic allocation
Move Octree to the base folder
Remove debug message prefix in vtk file
