+++ core/base/octree/CMakeLists.txt
@@ -1,0 +1,8 @@
+ttk_add_base_library(octree
+  SOURCES
+    Octree.cpp
+  HEADERS
+    Octree.h
+  DEPENDS
+    triangulation
+)
